+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * add default config options, making `DEBUG` and `TESTING` unnecessary to set manually
 * add a `_form_name` attribute to `FlaskForm` to automatically name forms when rendering them programmatically
 * add `get_extension_local_proxy` and `get_service_local_proxy` methods to the `Unchained` extension
+* add support for overriding static files from bundles
 
 ## 0.5.1 (2018/07/25)